requirements.txt
pyboy>=1.6.0
slack-bolt>=1.18.0
slack-sdk>=3.21.0
python-dotenv>=1.0.0
Pillow>=10.0.0
